Fix: RoomOverViewTableViewCell 오토레이아웃 수정
